fix: alter use of loadable-component
