+++ src/components/pages/experiments/scrollytelling/scrollytelling.module.scss
@@ -1,0 +1,11 @@
+.path {
+  stroke-dasharray: 1934.30322265625;
+  stroke-dashoffset: 1934.30322265625;
+  animation: dash 2s linear forwards;
+}
+
+@keyframes dash {
+  to {
+    stroke-dashoffset: 0;
+  }
+}
